+++ src/main/java/com/example/lit/controller/LitUpController.java
@@ -1,7 +1,6 @@
 package com.example.lit.controller;
 
-import com.example.lit.domain.vo.review.ReviewVO;
-import com.example.lit.service.LitUpService;
+import com.example.lit.service.review.LitUpService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -15,6 +14,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class LitUpController {
     private final LitUpService litUpService;
 
+    //테스트
+    @GetMapping("/write")
+    public String write(){
+        log.info("******************************");
+        log.info("LitUpController : write");
+        log.info("******************************");
+        return "/certification/certificationWrite";
+    }
 
+    //테스트
+    @GetMapping("/detail")
+    public String detail(){
+        log.info("******************************");
+        log.info("LitUpController : detail");
+        log.info("******************************");
+        return "/certification/certificationDetail";
+    }
 
 }
