--- src/main/java/com/example/lit/controller/LoginController.java
@@ -1,31 +1,0 @@
-package com.example.lit.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@Slf4j
-@RequestMapping("/login/*")
-@RequiredArgsConstructor
-public class LoginController {
-    //의존성 주입 객체 선언
-
-    @GetMapping("/login")
-    public String login(){
-        log.info("******************************");
-        log.info("LoginController : login");
-        log.info("******************************");
-        return "/login/login";
-    }
-
-    @GetMapping("/join")
-    public String join(){
-        log.info("******************************");
-        log.info("LoginController : join");
-        log.info("******************************");
-        return "/login/join";
-    }
-}
